backend/pyproject.toml



[tool.poetry]
name = "matcha_backend"
version = "0.1.0"
description = "matcha Backend"
authors = []
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]

python = "^3.11"
python-dotenv = "^1.0.0"
pydantic = { extras = ["email"], version = "^2.4.1" }
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
uvloop = "^0.17.0"
httptools = "^0.6.0"
uuid = "^1.30"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
python-jose = "^3.3.0"
pytest = "^7.4.2"
python-multipart = "^0.0.6"
httpx = "^0.25.0"
pydantic-settings = "^2.0.3"
gunicorn = "^21.2.0"
bcrypt = "^4.1.1"
aiohttp="^3.9.1"
requests="^2.31.0"
asyncio="^3.4.3"
pycocotools="^2.0.7"
scipy="^1.12.0"
orjson="^3.9.12"
psycopg2-binary = "^2.9.9"
jinja2="^3.1.3"
databases="^0.9.0"
pyjwt="^2.8.0"
dependency_injector="^4.42.0"
cloudinary="^1.41.0"
pydicom="^2.4.4"
pynetdicom = "^2.1.1"
dicomgenerator = "^0.8.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120
fix = true
select = [
    # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "F", # Pyflakes
    # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "E", # pycodestyle
    "W", # Warning
    # https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    # https://docs.astral.sh/ruff/rules/#mccabe-c90
    "C", # Complexity (mccabe+) & comprehensions
    # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "UP", # pyupgrade
    # https://docs.astral.sh/ruff/rules/#isort-i
    "I", # isort
]
ignore = [
    # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "E402", # module level import not at top of file
    # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "UP006", # use-pep585-annotation
    "UP007", # use-pep604-annotation
    "E741",  # Ambiguous variable name
    # "UP035", # deprecated-assertion
]
[tool.ruff.per-file-ignores]
"__init__.py" = [
    "F401", # unused import
    "F403", # star imports
]

[tool.ruff.mccabe]
max-complexity = 24

[tool.ruff.pydocstyle]
convention = "numpy"
